job-tracker-api/requirements.txt
# Production dependencies for Job Application Tracker API
# Updated: 2024-12-19

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client (if needed for external APIs)
httpx==0.25.2

# Date/Time handling
python-dateutil==2.8.2

# JSON handling
python-multipart==0.0.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing (minimal for production debugging)
pytest==7.4.3
pytest-asyncio==0.21.1

# Async support
asyncio-mqtt==0.16.1

# Type hints support
typing-extensions==4.8.0

# Production server
gunicorn==21.2.0
